+++ src/config/result/client_config.h
@@ -99,7 +99,7 @@ public:
     /// @param stringLimit maximum length of name strings
     /// @param multiValue client support multi value attributes
     /// @param isAllowed client is allowed to connect to server
-    ClientConfig(int flags, std::string_view group, std::string_view ip, std::string_view userAgent,
+    ClientConfig(QuirkFlags flags, std::string_view group, std::string_view ip, std::string_view userAgent,
         const std::map<ClientMatchType, std::string>& matchValues,
         int captionInfoCount, int stringLimit, bool multiValue, bool isAllowed);
 
@@ -109,8 +109,8 @@ public:
     /// @brief get client profile associated with configuration
     const ClientProfile& getClientProfile() const { return clientProfile; }
 
-    int getFlags() const { return this->clientProfile.flags; }
-    void setFlags(int flags) { this->clientProfile.flags = flags; }
+    QuirkFlags getFlags() const { return this->clientProfile.flags; }
+    void setFlags(QuirkFlags flags) { this->clientProfile.flags = flags; }
 
     /// @brief special mappings for client
     std::map<std::string, std::string> getMimeMappings(bool edit = false) const { return this->clientProfile.mimeMappings.getDictionaryOption(edit); }
@@ -193,13 +193,17 @@ public:
     static std::string_view mapClientType(ClientType clientType);
     static std::string_view mapMatchType(ClientMatchType matchType);
     static ClientMatchType remapMatchType(const std::string& matchType);
-    static int remapFlag(const std::string& flag);
-    static int makeFlags(const std::string& optValue);
+    static QuirkFlags remapFlag(const std::string& flag);
+    static QuirkFlags makeFlags(const std::string& optValue);
 
     static std::string mapFlags(QuirkFlags flags);
+    static bool hasFlag(QuirkFlags flag, Quirk quirkFlag);
+    static QuirkFlags getFlag(Quirk quirkFlag);
+    static QuirkFlags getFlags(const std::vector<Quirk> quirkFlags);
 
 protected:
     ClientProfile clientProfile;
+    static std::vector<std::pair<std::string_view, Quirk>> quirkFlags;
 };
 
 #endif