+++ src/request_handler/device_description_handler.cc
@@ -103,12 +103,12 @@ std::string DeviceDescriptionHandler::getPresentationUrl(const std::string& ip, 
 
 struct ServiceCapabilityValues {
     std::string value;
-    QuirkFlags quirkFlags;
+    Quirk quirkFlags;
 };
 
 struct ServiceCapabilities {
     std::string_view serviceType;
-    QuirkFlags quirkFlags;
+    Quirk quirkFlags;
     std::vector<ServiceCapabilityValues> serviceValues;
 };
 
@@ -155,29 +155,29 @@ std::string DeviceDescriptionHandler::renderDeviceDescription(const std::string&
     // add service details
     {
         const ServiceCapabilities deviceStringProperties[] = {
-            { "dlna:X_DLNACAP", QUIRK_FLAG_NONE, {} },
-            { "dlna:X_DLNADOC", QUIRK_FLAG_NONE, { { "DMS-1.50", QUIRK_FLAG_NONE } } },
-            { "dlna:X_DLNADOC", QUIRK_FLAG_SAMSUNG, { { "M-DMS-1.50", QUIRK_FLAG_NONE } } },
-            { "deviceType", QUIRK_FLAG_NONE, { { "urn:schemas-upnp-org:device:MediaServer:1", QUIRK_FLAG_NONE } } },
-            { "presentationURL", QUIRK_FLAG_NONE, { { getPresentationUrl(ip, port), QUIRK_FLAG_NONE } } },
-            { "sec:ProductCap", QUIRK_FLAG_NONE, {
-                                                     { "smi", QUIRK_FLAG_NONE },
-                                                     { "DCM10", QUIRK_FLAG_SAMSUNG },
-                                                     { "getMediaInfo.sec", QUIRK_FLAG_NONE },
-                                                     { "getCaptionInfo.sec", QUIRK_FLAG_NONE },
-                                                 } },
-            { "sec:X_ProductCap", QUIRK_FLAG_SAMSUNG, {
-                                                          { "smi", QUIRK_FLAG_NONE },
-                                                          { "DCM10", QUIRK_FLAG_NONE },
-                                                          { "getMediaInfo.sec", QUIRK_FLAG_NONE },
-                                                          { "getCaptionInfo.sec", QUIRK_FLAG_NONE },
-                                                      } },
+            { "dlna:X_DLNACAP", Quirk::None, {} },
+            { "dlna:X_DLNADOC", Quirk::None, { { "DMS-1.50", Quirk::None } } },
+            { "dlna:X_DLNADOC", Quirk::Samsung, { { "M-DMS-1.50", Quirk::None } } },
+            { "deviceType", Quirk::None, { { "urn:schemas-upnp-org:device:MediaServer:1", Quirk::None } } },
+            { "presentationURL", Quirk::None, { { getPresentationUrl(ip, port), Quirk::None } } },
+            { "sec:ProductCap", Quirk::None, {
+                                                 { "smi", Quirk::None },
+                                                 { "DCM10", Quirk::Samsung },
+                                                 { "getMediaInfo.sec", Quirk::None },
+                                                 { "getCaptionInfo.sec", Quirk::None },
+                                             } },
+            { "sec:X_ProductCap", Quirk::Samsung, {
+                                                      { "smi", Quirk::None },
+                                                      { "DCM10", Quirk::None },
+                                                      { "getMediaInfo.sec", Quirk::None },
+                                                      { "getCaptionInfo.sec", Quirk::None },
+                                                  } },
         };
         for (auto&& [tag, serviceFlags, values] : deviceStringProperties) {
-            if (!quirks || serviceFlags == QUIRK_FLAG_NONE || quirks->hasFlag(serviceFlags)) {
+            if (!quirks || serviceFlags == Quirk::None || quirks->hasFlag(serviceFlags)) {
                 std::vector<std::string> serviceValue;
                 for (auto&& [value, valueFlags] : values) {
-                    if (!quirks || valueFlags == QUIRK_FLAG_NONE || quirks->hasFlag(valueFlags)) {
+                    if (!quirks || valueFlags == Quirk::None || quirks->hasFlag(valueFlags)) {
                         serviceValue.push_back(value);
                     }
                 }